backend/requirements.txt
Flask
flask-cors
numpy
pillow
opencv-python
torch
torchvision
imagehash
